+++ apps/backend/src/modules/pathway/dtos/update-participant-status.dto.ts
@@ -1,16 +1,16 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IsIn, IsNotEmpty, IsString } from 'class-validator';
+import { IsEnum, IsNotEmpty } from 'class-validator';
+import { PathwayParticipantStatusEnum } from '@src/commons/enums';
 
 export class UpdateParticipantStatusDto {
   @ApiProperty({
     example: 'in_progress',
     description: 'Participant status',
-    enum: ['invited', 'in_progress', 'completed'],
+    enum: PathwayParticipantStatusEnum,
   })
   @IsNotEmpty({ message: 'Status is required' })
-  @IsString()
-  @IsIn(['invited', 'in_progress', 'completed'], {
+  @IsEnum(PathwayParticipantStatusEnum, {
     message: 'Status must be invited, in_progress, or completed',
   })
-  status!: string;
+  status!: PathwayParticipantStatusEnum;
 }
